Big controls update - add update messages - remove controls map (no need on this side) - move controls out of its own folder
This update fleshes out the update method in controls and ups the
frequency to 20Hz. The controls map was removed because I decided
against having any control logic running on the operator console in
favor of having it all on the rover's main computer. Buttons will no
longer be sent as diffs in the interest of increasing resiliency to
network interruptions. We don't want to get the rover stuck running full
tilt towards a cliff when we losr connection.
Now that the map.json file no longer exists, the Controls.ts file was in
its own folder, so I moved it up to the root of the ts folder.

[#249] Get basic joystick control hooked in

https://project.red2fred2.com/work_packages/249